+++ client/src/androidUnitTest/kotlin/org/apptank/horus/client/control/SyncControlDatabaseHelperTest.kt
@@ -6,6 +6,10 @@ import org.apptank.horus.client.TestCase
 import org.apptank.horus.client.control.scheme.SyncControlTable
 import org.apptank.horus.client.data.Horus
 import org.apptank.horus.client.cache.MemoryCache
+import org.apptank.horus.client.control.scheme.DataSharedTable
+import org.apptank.horus.client.control.scheme.EntitiesTable
+import org.apptank.horus.client.control.scheme.EntityAttributesTable
+import org.apptank.horus.client.control.scheme.SyncFileTable
 import org.apptank.horus.client.database.HorusDatabase
 import org.apptank.horus.client.database.SyncControlDatabaseHelper
 import org.apptank.horus.client.extensions.execute
@@ -38,6 +42,10 @@ class SyncControlDatabaseHelperTest : TestCase() {
 
         Assert.assertTrue(tablesNames.contains(SyncControlTable.TABLE_NAME))
         Assert.assertTrue(tablesNames.contains(QueueActionsTable.TABLE_NAME))
+        Assert.assertTrue(tablesNames.contains(EntitiesTable.TABLE_NAME))
+        Assert.assertTrue(tablesNames.contains(EntityAttributesTable.TABLE_NAME))
+        Assert.assertTrue(tablesNames.contains(SyncFileTable.TABLE_NAME))
+        Assert.assertTrue(tablesNames.contains(DataSharedTable.TABLE_NAME))
     }
 
     @Test
@@ -433,6 +441,26 @@ class SyncControlDatabaseHelperTest : TestCase() {
 
         // Then
         Assert.assertEquals(0, controlManagerDatabaseHelper.getTables().size)
+    }
+
+    @Test
+    fun validateGetReadableEntitiesIsSuccess() {
+        // Given
+        val entityWritable = "entity_writable_123"
+        val entityReadOnly = "entity_read_only_423"
+
+        driver.createTable(entityWritable, mapOf("id" to "TEXT"))
+        driver.createTable(entityReadOnly, mapOf("id" to "TEXT"))
+
+        driver.registerEntity(entityWritable)
+        driver.registerEntity(entityReadOnly, false)
+
+        // When
+        val readableEntities = controlManagerDatabaseHelper.getReadableEntityNames()
+
+        // Then
+        Assert.assertEquals(1, readableEntities.size)
+        assert(readableEntities.contains(entityReadOnly))
     }
 
 }