+++ client/src/commonMain/kotlin/org/apptank/horus/client/control/helper/IOperationDatabaseHelper.kt
@@ -69,12 +69,14 @@ interface IOperationDatabaseHelper {
      * @param table The name of the table.
      * @param conditions The list of conditions for deletion.
      * @param operator The logical operator to combine conditions (AND/OR).
+     * @param disableForeignKeys Flag to disable foreign key checks during deletion.
      * @return The result of the operation.
      */
     fun deleteRecords(
         table: String,
         conditions: List<SQL.WhereCondition>,
-        operator: SQL.LogicOperator = SQL.LogicOperator.AND
+        operator: SQL.LogicOperator = SQL.LogicOperator.AND,
+        disableForeignKeys: Boolean = false
     ): DatabaseOperation.Result
 
     /**
@@ -93,4 +95,13 @@ interface IOperationDatabaseHelper {
      * @return the count of records from the query result.
      */
     fun countRecords(builder: SimpleQueryBuilder): Int
+
+
+    /**
+     * Truncates the specified entity from the database.
+     *
+     * @param entity The name of the entity to be truncated.
+     * @return True if the truncation was successful, false otherwise.
+     */
+    fun truncate(entity: String)
 }