+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# v0.11.0
+- Added support to shared entities.
+- Fixed issue when is delete corrupted data by foreign key constraint.
+- Added refresh readable data entities.
+
 ## v0.10.3
 - Fixed filter delete operations in batch.